rename Pizza Componant  file
